some more code cleaning and file restructuring.
