+++ Code/Localr/navigation/AuthNavigator.js
@@ -1,7 +1,10 @@
 import React, { useState, useEffect, createContext } from 'react'
 import Firebase from '../components/Firebase'
-import SignInStack from './SignInStack'
+import SignInStack from './SignInDrawer'
 import SignOutStack from './SignOutStack'
+
+//read this to understand what is happening, there are two navigation containers depending if you are logged in or not, this file chooses which container to use.
+//https://heartbeat.fritz.ai/how-to-manage-authentication-flows-in-react-native-with-react-navigation-v5-and-firebase-860f57ae20d3
 
 export const AuthContext = createContext(null)
 
@@ -28,7 +31,7 @@ export default function AuthNavigator() {
 
     return user ? (
       <AuthContext.Provider value={user}>
-        <SignInStack />
+        <SignInDrawer/>
       </AuthContext.Provider>
     ) : (
       <SignOutStack />
